Move StrategyContextMissing to exceptions module and add test fixture
