+++ backend/core/logic/letters/exceptions.py
@@ -1,0 +1,4 @@
+class StrategyContextMissing(Exception):
+    """Raised when required strategy context is missing for finalization."""
+    pass
+
