+++ tests/fixtures/Experian_gpt_response.json
@@ -1,0 +1,5 @@
+{
+  "accounts": [],
+  "meta": {"bureau": "Experian", "generated": "2025-01-01T00:00:00Z"}
+}
+
